refactor : PR 의견을 반영하여 auth.domain 내 객체를 core 모듈로 이동 / style : code refactor (+ optimizer imports)
