+++ README.md
@@ -22,6 +22,6 @@
 
 |            [gentledot](https://github.com/GentleDot)             |            [highright96](https://github.com/highright96)             |  
 | :----------------------------------------------------------: | :----------------------------------------------------------: 
-| <img src="https://user-images.githubusercontent.com/55661631/145813189-67f4b845-a9f7-490e-837b-5f7997305f27.png" width=200px alt="_"/> | <img src="https://user-images.githubusercontent.com/55661631/141674899-d7496769-6736-47ee-a0cc-2c631520790a.jpg" width=200px alt="_"/> 
+| <img src="https://user-images.githubusercontent.com/55661631/145813189-67f4b845-a9f7-490e-837b-5f7997305f27.png" width=200px alt="_"/> | <img src="https://user-images.githubusercontent.com/55661631/141674899-d7496769-6736-47ee-a0cc-2c631520790a.jpg" width=200px alt="_"/>
 |                         백엔드                         |                        백엔드                         |
