+++ api/src/main/java/com/bookwhale/auth/controller/OAuthController.java
@@ -1,9 +1,9 @@
 package com.bookwhale.auth.controller;
 
 import com.bookwhale.auth.domain.CurrentUser;
-import com.bookwhale.auth.domain.provider.OAuthProviderType;
 import com.bookwhale.auth.dto.OAuthLoginResponse;
 import com.bookwhale.auth.dto.OAuthRefreshLoginRequest;
+import com.bookwhale.auth.dto.provider.OAuthProviderType;
 import com.bookwhale.auth.service.OauthService;
 import com.bookwhale.user.domain.User;
 import javax.validation.Valid;
