+++ api/src/main/java/com/bookwhale/auth/dto/OAuthObjectConverter.java
@@ -1,20 +1,20 @@
-package com.bookwhale.auth.domain;
+package com.bookwhale.auth.dto;
 
+import com.bookwhale.auth.domain.JWT;
+import com.bookwhale.auth.domain.JWT.Claims;
+import com.bookwhale.auth.domain.JWT.ClaimsForRefresh;
 import com.bookwhale.auth.domain.info.UserInfo;
 import com.bookwhale.auth.domain.info.UserInfoFromGoogle;
 import com.bookwhale.auth.domain.info.UserInfoFromNaver;
-import com.bookwhale.auth.domain.provider.GoogleOAuthProvider;
-import com.bookwhale.auth.domain.provider.NaverOAuthProvider;
-import com.bookwhale.auth.domain.provider.OAuthProvider;
-import com.bookwhale.auth.domain.provider.OAuthProviderType;
 import com.bookwhale.auth.domain.token.GoogleOAuthToken;
 import com.bookwhale.auth.domain.token.NaverOAuthToken;
 import com.bookwhale.auth.domain.token.OAuthToken;
+import com.bookwhale.auth.dto.provider.GoogleOAuthProvider;
+import com.bookwhale.auth.dto.provider.NaverOAuthProvider;
+import com.bookwhale.auth.dto.provider.OAuthProvider;
+import com.bookwhale.auth.dto.provider.OAuthProviderType;
 import com.bookwhale.common.exception.CustomException;
 import com.bookwhale.common.exception.ErrorCode;
-import com.bookwhale.common.token.JWT;
-import com.bookwhale.common.token.JWT.Claims;
-import com.bookwhale.common.token.JWT.ClaimsForRefresh;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.LinkedHashMap;
