+++ api/src/main/java/com/bookwhale/auth/service/OauthService.java
@@ -1,20 +1,20 @@
 package com.bookwhale.auth.service;
 
-import com.bookwhale.auth.domain.OAuthObjectConverter;
+import com.bookwhale.auth.domain.JWT;
+import com.bookwhale.auth.domain.JWT.Claims;
+import com.bookwhale.auth.domain.JWT.ClaimsForRefresh;
 import com.bookwhale.auth.domain.Token;
 import com.bookwhale.auth.domain.TokenRepository;
 import com.bookwhale.auth.domain.info.UserInfo;
-import com.bookwhale.auth.domain.provider.GoogleOAuthProvider;
-import com.bookwhale.auth.domain.provider.NaverOAuthProvider;
-import com.bookwhale.auth.domain.provider.OAuthProvider;
-import com.bookwhale.auth.domain.provider.OAuthProviderType;
 import com.bookwhale.auth.dto.OAuthLoginResponse;
+import com.bookwhale.auth.dto.OAuthObjectConverter;
 import com.bookwhale.auth.dto.OAuthRefreshLoginRequest;
+import com.bookwhale.auth.dto.provider.GoogleOAuthProvider;
+import com.bookwhale.auth.dto.provider.NaverOAuthProvider;
+import com.bookwhale.auth.dto.provider.OAuthProvider;
+import com.bookwhale.auth.dto.provider.OAuthProviderType;
 import com.bookwhale.common.exception.CustomException;
 import com.bookwhale.common.exception.ErrorCode;
-import com.bookwhale.common.token.JWT;
-import com.bookwhale.common.token.JWT.Claims;
-import com.bookwhale.common.token.JWT.ClaimsForRefresh;
 import com.bookwhale.common.utils.RandomUtils;
 import com.bookwhale.user.domain.User;
 import com.bookwhale.user.service.UserService;
