+++ api/src/test/java/com/bookwhale/chatroom/acceptance/ChatRoomAcceptanceTest.java
@@ -51,8 +51,10 @@ public class ChatRoomAcceptanceTest extends AcceptanceTest {
     @DisplayName("채팅방을 생성한다.")
     @Test
     void createChatRoom() {
-        String loginUserJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(user), jwt);
-        String sellerJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(anotherUser), jwt);
+        String loginUserJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(user), jwt);
+        String sellerJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(anotherUser), jwt);
 
         ExtractableResponse<Response> res = ChatRoomAcceptanceStep.createChatRoom(loginUserJwt,
             sellerJwt, articleRequest);
@@ -63,8 +65,10 @@ public class ChatRoomAcceptanceTest extends AcceptanceTest {
     @DisplayName("채팅방들을 조회한다.")
     @Test
     void findChatRooms() {
-        String loginUserJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(user), jwt);
-        String sellerJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(anotherUser), jwt);
+        String loginUserJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(user), jwt);
+        String sellerJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(anotherUser), jwt);
 
         ChatRoomAcceptanceStep.createChatRoom(loginUserJwt, sellerJwt, articleRequest);
 
@@ -89,8 +93,10 @@ public class ChatRoomAcceptanceTest extends AcceptanceTest {
     @DisplayName("상대방이 나간 채팅방은 isOpponentDelete 가 false 로 조회된다.")
     @Test
     void findChatRooms_opponentDelete() {
-        String loginUserJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(user), jwt);
-        String sellerJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(anotherUser), jwt);
+        String loginUserJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(user), jwt);
+        String sellerJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(anotherUser), jwt);
 
         ChatRoomAcceptanceStep.createChatRoom(loginUserJwt, sellerJwt, articleRequest);
 
@@ -114,8 +120,10 @@ public class ChatRoomAcceptanceTest extends AcceptanceTest {
     @DisplayName("나간 채팅방은 조회되지 않는다.")
     @Test
     void findChatRooms_deleteRoom() {
-        String loginUserJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(user), jwt);
-        String sellerJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(anotherUser), jwt);
+        String loginUserJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(user), jwt);
+        String sellerJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(anotherUser), jwt);
 
         ChatRoomAcceptanceStep.createChatRoom(loginUserJwt, sellerJwt, articleRequest);
 
@@ -136,8 +144,10 @@ public class ChatRoomAcceptanceTest extends AcceptanceTest {
     @DisplayName("채팅방을 삭제한 후 조회한다.")
     @Test
     void deleteChatRoom() {
-        String loginUserJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(user), jwt);
-        String sellerJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(anotherUser), jwt);
+        String loginUserJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(user), jwt);
+        String sellerJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(anotherUser), jwt);
 
         ChatRoomAcceptanceStep.createChatRoom(loginUserJwt, sellerJwt, articleRequest);
         Long roomId = ChatRoomAcceptanceStep.requestToFindChatRooms(loginUserJwt).jsonPath()
