+++ api/src/test/java/com/bookwhale/user/acceptance/UserAcceptanceTest.java
@@ -13,10 +13,7 @@ import com.bookwhale.common.acceptance.step.AcceptanceStep;
 import com.bookwhale.user.acceptance.step.UserAcceptanceStep;
 import com.bookwhale.user.dto.FavoriteRequest;
 import com.bookwhale.user.dto.FavoriteResponse;
-import com.bookwhale.user.dto.LoginRequest;
-import com.bookwhale.user.dto.PasswordUpdateRequest;
 import com.bookwhale.user.dto.ProfileResponse;
-import com.bookwhale.user.dto.SignUpRequest;
 import com.bookwhale.user.dto.UserResponse;
 import com.bookwhale.user.dto.UserUpdateRequest;
 import io.restassured.builder.MultiPartSpecBuilder;
@@ -51,7 +48,8 @@ public class UserAcceptanceTest extends AcceptanceTest {
             .nickname("hose12")
             .build();
 
-        String apiToken = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(user), jwt);
+        String apiToken = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(user), jwt);
 
         ExtractableResponse<Response> response = UserAcceptanceStep.requestToUpdateMyInfo(apiToken,
             userUpdateRequest);
@@ -72,9 +70,11 @@ public class UserAcceptanceTest extends AcceptanceTest {
             .fileName("profileImage.jpg")
             .build();
 
-        String apiToken = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(user), jwt);
+        String apiToken = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(user), jwt);
 
-        ExtractableResponse<Response> response = UserAcceptanceStep.uploadProfileImage(apiToken, image);
+        ExtractableResponse<Response> response = UserAcceptanceStep.uploadProfileImage(apiToken,
+            image);
         ProfileResponse profileResponse = response.jsonPath().getObject(".", ProfileResponse.class);
         UserResponse userResponse = UserAcceptanceStep.requestToGetMyInfo(apiToken).jsonPath()
             .getObject(".", UserResponse.class);
@@ -93,7 +93,8 @@ public class UserAcceptanceTest extends AcceptanceTest {
             .fileName("profileImage.jpg")
             .build();
 
-        String apiToken = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(user), jwt);
+        String apiToken = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(user), jwt);
 
         UserAcceptanceStep.uploadProfileImage(apiToken, image);
         ExtractableResponse<Response> response = UserAcceptanceStep.deleteProfileImage(apiToken);
@@ -125,13 +126,15 @@ public class UserAcceptanceTest extends AcceptanceTest {
             .price("5000")
             .build();
 
-        String apiToken = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(user), jwt);
+        String apiToken = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(user), jwt);
         Long articleId = AcceptanceUtils.getIdFromResponse(
             ArticleAcceptanceStep.requestToCreateArticle(apiToken, articleRequest));
 
         ExtractableResponse<Response> response = UserAcceptanceStep.addFavorite(apiToken,
             new FavoriteRequest(articleId));
-        List<FavoriteResponse> favoriteRespons = UserAcceptanceStep.findFavorites(apiToken).jsonPath()
+        List<FavoriteResponse> favoriteRespons = UserAcceptanceStep.findFavorites(apiToken)
+            .jsonPath()
             .getList(".", FavoriteResponse.class);
 
         AcceptanceStep.assertThatStatusIsOk(response);
@@ -159,7 +162,8 @@ public class UserAcceptanceTest extends AcceptanceTest {
             .price("5000")
             .build();
 
-        String apiToken = UserAcceptanceStep.requestToLoginAndGetAccessToken(UserInfoFromToken.of(user), jwt);
+        String apiToken = UserAcceptanceStep.requestToLoginAndGetAccessToken(
+            UserInfoFromToken.of(user), jwt);
         Long articleId = AcceptanceUtils.getIdFromResponse(
             ArticleAcceptanceStep.requestToCreateArticle(apiToken, articleRequest));
         UserAcceptanceStep.addFavorite(apiToken, new FavoriteRequest(articleId));
@@ -172,7 +176,8 @@ public class UserAcceptanceTest extends AcceptanceTest {
 
         ExtractableResponse<Response> response = UserAcceptanceStep.deleteFavorite(
             apiToken, favoriteId);
-        List<FavoriteResponse> favoriteRespons = UserAcceptanceStep.findFavorites(apiToken).jsonPath()
+        List<FavoriteResponse> favoriteRespons = UserAcceptanceStep.findFavorites(apiToken)
+            .jsonPath()
             .getList(".", FavoriteResponse.class);
 
         ExtractableResponse<Response> responseAfterDeleteFavorite = ArticleAcceptanceStep.requestToFindArticle(
