+++ api/src/test/java/com/bookwhale/user/acceptance/step/UserAcceptanceStep.java
@@ -4,19 +4,15 @@ import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.bookwhale.article.dto.ArticleRequest;
-import com.bookwhale.auth.domain.OAuthObjectConverter;
+import com.bookwhale.auth.domain.JWT;
 import com.bookwhale.auth.domain.info.UserInfo;
-import com.bookwhale.auth.domain.info.UserInfoFromToken;
-import com.bookwhale.common.token.JWT;
-import com.bookwhale.common.token.JWT.Claims;
+import com.bookwhale.auth.dto.OAuthObjectConverter;
 import com.bookwhale.user.domain.User;
 import com.bookwhale.user.dto.FavoriteRequest;
 import com.bookwhale.user.dto.FavoriteResponse;
 import com.bookwhale.user.dto.LoginRequest;
-import com.bookwhale.user.dto.LoginResponse;
 import com.bookwhale.user.dto.PasswordUpdateRequest;
 import com.bookwhale.user.dto.ProfileResponse;
-import com.bookwhale.user.dto.SignUpRequest;
 import com.bookwhale.user.dto.UserResponse;
 import com.bookwhale.user.dto.UserUpdateRequest;
 import io.restassured.response.ExtractableResponse;
