+++ core/build.gradle
@@ -9,8 +9,18 @@ dependencies {
     //database
     runtimeOnly 'com.h2database:h2'
     runtimeOnly 'mysql:mysql-connector-java'
+
+    //auth
+    implementation 'io.jsonwebtoken:jjwt:0.5.1'
+    implementation 'com.auth0:java-jwt:3.18.2'
+
+    // util
+    implementation 'org.apache.commons:commons-lang3:3.12.0'
+    implementation 'com.google.guava:guava:31.0.1-jre'
+
     implementation 'com.querydsl:querydsl-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+
 }
 
 ext {
