+++ core/src/main/java/com/bookwhale/auth/domain/CurrentUser.java
