+++ core/src/main/java/com/bookwhale/auth/domain/info/UserInfo.java
