+++ core/src/main/java/com/bookwhale/auth/domain/info/UserInfoFromGoogle.java
