+++ core/src/main/java/com/bookwhale/auth/domain/info/UserInfoFromNaver.java
