+++ core/src/main/java/com/bookwhale/auth/domain/info/UserInfoFromToken.java
