+++ core/src/main/java/com/bookwhale/auth/domain/token/NaverOAuthToken.java
